requirements.txt
advertorch==0.2.3
boto3==1.22.7
botocore==1.25.7
colorama==0.4.4
falconn==1.3.1
matplotlib==3.5.1
numba==0.54.1
numpy==1.20.3
pandas==1.2.5
Pillow==9.1.0
scipy==1.7.3
seaborn==0.11.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
